requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic>=2.0
